pyproject.toml
[tool.poetry]
name = "foodgram"
version = "0.1.0"
description = ""
authors = ["Konstantin Raikhert <konstantinraihert@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
django-filter = "^2.4.0"
djangorestframework-simplejwt = "^4.7.2"
Pillow = "^8.3.1"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.1"
python-environ = "^0.4.54"
django-cors-headers = "^3.7.0"
djangorestframework-camel-case = "^1.2.0"
factory-boy = "^3.2.0"
djoser = "^2.1.0"

[tool.poetry.dev-dependencies]
pylint = "^2.9.4"
pre-commit = "^2.13.0"
black = "^21.7b0"
flake8 = "^3.9.2"
isort = "^5.9.2"
django-extensions = "^3.1.3"
django-debug-toolbar = "^3.2.1"
pytest-django = "^4.4.0"
django-coverage-plugin = "^2.0.0"
ipython = "^7.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.pyright]
include = ["backend"]
